requirements.txt
streamlit
google-genai
pandas
PyPDF2
google-auth-oauthlib
google-api-python-client
python-dotenv
python-docx
python-pptx
